requirement.txt
cffi==1.17.1
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
fonttools==4.55.0
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pycparser==2.22
pycryptodome==3.21.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
seaborn==0.13.2
six==1.16.0
tzdata==2024.2
pycryptodome==3.21.0
